+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * caution text to system YELLOW
 * adb print cleanup
 * mediaRecordPlacebo to VOICE_COMM source
+* browser intent ex null
 
 ### 4.4.0 (2020-6-27)
 * bugfix - add context for entryLogger